requirements.txt
agno 
pandas 
openpyxl 
PyPDF2 
python-docx 
fpdf
pdfplumber
reportlab
openai
